Buck-Boost converter .tex
Added the .tex file with the background and introduction to both the unidirectional and bidirectional BuckBoost converter. Also I added 2 images with the schematics. However I'm not sure about the folders where I put everything. Plus, the document I created is set as a standalone file. If the document organizer tells me the name of my section I'll do the appropriate changes